requirements.txt
streamlit==1.36.0
uvicorn==0.30.1
fastapi==0.111.1
keras==3.4.1
pandas==2.0.3
plotly==5.23.0
yfinance==0.2.41
numpy==1.26.4
pydantic==2.2.1
tensorflow==2.17.0
scikit-learn==1.5.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0